refactor: move translateSeverityToExitCode from diagnostics to orchestrator
