+++ src/orchestrator/index.ts
@@ -4,3 +4,4 @@
 
 export { inspectFiles, inspectProject } from "./executor.ts";
 export type { InspectOptions } from "./options.ts";
+export { translateSeverityToExitCode } from "./tools.ts";
